# Accelerate CUDA Top 1 Select