Previous implementation did not parallelize across logits in a batch. Average performance measured on RTX 4090 over 1000 iterations using vocab size 32000.

| Average performance per batch size | Cuda Argmax (unbatched) | Custom Sampling Kernel (batched) |
|--------------------|---------------|--------------|
| Batch Size 1          | 24.62μs          | 22.86μs        |
| Batch Size 5       | 63.30μs         | 25.55μs        |
| Batch Size 12          | 120.35μs           | 23.6μs          |